test(ISQLStatementGenerator): Add test case for generateCreateTableStatements
